Added Files to Project for GDIT
Hi Terri,

I must say I enjoyed this assignment, especially after a LONG day of packing for a move.
I have the following files attached with detail pertaining to what each includes:

ReadMe.txt- This file includes my answers to the 3 questions listed in attachment 1 in your initial email. There is also detail about the other files I am listing below.
TestPlan.txt- This file details my test plan as far as what is in scope and what was out of scope of my testing. It also includes my test cases in Gherkin/BDD syntax so that the tests are reader-friendly.
PseudoCode.txt- This file is my pseudocode for my automated test cases.
GDITLoginPOM.py- This file includes some rough python code that is the POM file for Login Testing-- this is to create reusability throughout testing
GDITForgotPasswordRecover.py- This file as well includes rough python code that is the POM file for Password recovery, once a user selects "forgot password", as it is assumed this is a new page-- This as well creates reusability and encapsulation